+++ TractorMod/Framework/Attachments/PickaxeAttachment.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+using System.Linq;
 using Microsoft.Xna.Framework;
 using StardewValley;
 using StardewValley.TerrainFeatures;
@@ -21,6 +23,14 @@ namespace Pathoschild.Stardew.TractorMod.Framework.Attachments
 
         /// <summary>Whether to clear tilled dirt.</summary>
         private readonly bool ClearDirt;
+
+        /// <summary>The axe upgrade levels needed to break supported resource clumps.</summary>
+        /// <remarks>Derived from <see cref="ResourceClump.performToolAction"/>.</remarks>
+        private readonly IDictionary<int, int> ResourceUpgradeLevelsNeeded = new Dictionary<int, int>
+        {
+            [ResourceClump.meteoriteIndex] = Tool.gold,
+            [ResourceClump.boulderIndex] = Tool.steel
+        };
 
 
         /*********
@@ -73,6 +83,26 @@ namespace Pathoschild.Stardew.TractorMod.Framework.Attachments
                 // clear dead crops
                 if (dirt.crop?.dead == true)
                     return this.UseToolOnTile(tool, tile);
+            }
+
+            // clear boulders / meteorites
+            // This needs to check if the axe upgrade level is high enough first, to avoid spamming
+            // 'need to upgrade your tool' messages. Based on ResourceClump.performToolAction.
+            if (this.BreakRocks)
+            {
+                Rectangle tileArea = this.GetAbsoluteTileArea(tile);
+                ResourceClump stump =
+                    (
+                        from clump in this.GetResourceClumps(location)
+                        where
+                        clump.getBoundingBox(clump.tile).Intersects(tileArea)
+                        && this.ResourceUpgradeLevelsNeeded.ContainsKey(clump.parentSheetIndex)
+                        && tool.upgradeLevel >= this.ResourceUpgradeLevelsNeeded[clump.parentSheetIndex]
+                        select clump
+                    )
+                    .FirstOrDefault();
+                if (stump != null)
+                    this.UseToolOnTile(tool, tile);
             }
 
             return false;